NavRail Component 분리
